deprecate HTML rendering of results
This commit deprecates the functionality in this project that renders
HUD results in an HTML page. This functionality previously relied on an
external template file hud_list.html which lives in CFPB's cfgov-refresh
repository. This made it impossible to run using this repository alone.

For that reason, and to simplify this repository into an API that only
provides data in JSON (and CSV) format, the HTML rendering functionality
has been removed.